+++ src/types/card.ts
@@ -1,0 +1,8 @@
+export type Card = {
+  id: string;
+  cardNumber: string;
+  cardBrand: string;
+  expireDate: string;
+  cvc: string;
+  cardPassword: string;
+};
